Changing location of a constant
